requirements.txt
rdkit-pypi
pandas
numpy
scikit-learn
lightgbm
xgboost
optuna
mordred
shap
fastapi
uvicorn
joblib
pydantic
pyyaml
requests
